requirements.txt
numpy
wandb
Pillow
datetime
matplotlib
torchvision
cupy-cuda101
